Changes to error types and moved some functions into modules
