+++ src/v4/aes.rs
@@ -1,10 +1,18 @@
 // This file contains things related to the V4 AES edek, which is defined in the icl_v4_header.proto.
 
 use crate::{
-    aes::{aes_decrypt_core, aes_encrypt, EncryptionKey},
-    create_signed_proto, icl_header_v4, Error,
+    aes::{self, aes_decrypt_core, aes_encrypt, EncryptionKey},
+    icl_header_v4::{
+        self,
+        v4document_header::{
+            signature_information::SignatureType, SignatureInformation, SignedPayload,
+        },
+        V4DocumentHeader,
+    },
+    signing, Error,
 };
 use bytes::Bytes;
+use protobuf::Message;
 use rand::{CryptoRng, RngCore};
 
 type Result<T> = core::result::Result<T, crate::Error>;
@@ -34,6 +42,42 @@ pub fn generate_aes_edek<R: CryptoRng + RngCore>(
         ..Default::default()
     };
     Ok((dek, aes_edek))
+}
+
+/// Sign the payload using the key.
+pub fn sign_header(
+    key: aes::EncryptionKey,
+    header_payload: &SignedPayload,
+) -> SignatureInformation {
+    //This unwrap can't actually ever happen because they create the coded stream with exactly the computed size before
+    //serializing.
+    let bytes = header_payload
+        .write_to_bytes()
+        .expect("Writing proto to bytes failed.");
+    let signature = signing::sign_hs256(key.0, &bytes);
+
+    SignatureInformation {
+        signature: signature.0.to_vec().into(),
+        signature_type: SignatureType::HS256.into(),
+        ..Default::default()
+    }
+}
+
+/// Creates a signed proto wrapper with a single edek wrapper in it using the signing key to do the signing.
+pub fn create_signed_proto(
+    edek_wrappers: Vec<icl_header_v4::v4document_header::EdekWrapper>,
+    signing_key: aes::EncryptionKey,
+) -> V4DocumentHeader {
+    let signed_payload = icl_header_v4::v4document_header::SignedPayload {
+        edeks: edek_wrappers,
+        ..Default::default()
+    };
+    let signature_info = sign_header(signing_key, &signed_payload);
+    icl_header_v4::V4DocumentHeader {
+        signed_payload: Some(signed_payload).into(),
+        signature_info: Some(signature_info).into(),
+        ..Default::default()
+    }
 }
 
 /// If `maybe_dek` is None, generate a dek, otherwise use the one provided.
@@ -79,10 +123,40 @@ pub fn decrypt_aes_edek(
         .map(EncryptionKey)
 }
 
+/// Verify the signature inside the V4 header
+pub fn verify_signature(key: aes::EncryptionKey, header: &V4DocumentHeader) -> bool {
+    match header.signature_info.signature_type.enum_value() {
+        Ok(SignatureType::NONE) => true,
+        Ok(SignatureType::HS256) => {
+            if let Ok(signature_bytes) = header.signature_info.signature.to_vec().try_into() {
+                signing::verify_hs256(
+                    key.0,
+                    //This unwrap can't actually ever happen because they create the coded stream with exactly the computed size before
+                    //serializing.
+                    &header
+                        .signed_payload
+                        .write_to_bytes()
+                        .expect("Writing proto to bytes failed."),
+                    &signing::Signature(signature_bytes),
+                )
+            } else {
+                false
+            }
+        }
+        _ => false,
+    }
+}
+
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::verify_signature;
+    use crate::{
+        icl_header_v4::v4document_header::{
+            edek_wrapper::{Aes256GcmEncryptedDek, Edek},
+            EdekWrapper,
+        },
+        signing::AES_KEY_LEN,
+    };
     use hex_literal::hex;
     use protobuf::Message;
     use rand::SeedableRng;
@@ -144,5 +218,42 @@ mod test {
         let verify_result = verify_signature(decrypted_aes_dek, &v4_document);
         // Verify fails because I messed the signature up in the proto_bytes
         assert!(!verify_result)
+    }
+
+    #[test]
+    fn sign_verify_roundtrip() {
+        let dek = EncryptionKey([100u8; AES_KEY_LEN]);
+        let aes_edek = Aes256GcmEncryptedDek {
+            ciphertext: [42u8; 1024].as_ref().into(),
+            ..Default::default()
+        };
+
+        let edek_wrapper = EdekWrapper {
+            edek: Some(Edek::Aes256GcmEdek(aes_edek)),
+            ..Default::default()
+        };
+
+        let signed_payload = SignedPayload {
+            edeks: vec![edek_wrapper],
+            ..Default::default()
+        };
+
+        let mut header = V4DocumentHeader {
+            signed_payload: Some(signed_payload).into(),
+            ..Default::default()
+        };
+
+        let sign_result = sign_header(dek, &header.signed_payload);
+
+        header.signature_info = Some(sign_result).into();
+        assert!(verify_signature(dek, &header));
+    }
+
+    #[test]
+    fn verify_known_good_sig_in_v4_header() {
+        let dek = EncryptionKey([100u8; AES_KEY_LEN]);
+        let bytes = hex_literal::hex!("0a240a2082e7f2abc390635636f59ea51f7736846d9b1e799f4e9b63733679a417a2c5cf10011289081286081a83081280082a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a2a");
+        let header = Message::parse_from_bytes(&bytes).unwrap();
+        assert!(verify_signature(dek, &header))
     }
 }
